[fix][x86] fix int datatype for some kernel

+++ src/ppl/nn/engines/x86/impls/src/ppl/kernel/x86/fp32/arithmetic/sse/arithmetic_broadcast_ndarray_fp32_sse.h
@@ -307,7 +307,7 @@ static ppl::common::RetCode arithmetic_broadcast_ndarray_fp32_sse(
         idx2dims(start_idx, real_dst_shape, real_dim_count, blocks[i].start);
         idx2dims(end_idx, real_dst_shape, real_dim_count, blocks[i].end);
         blocks[i].id = i;
-        for (int j = 0; j < real_dim_count; j++) {
+        for (int64_t j = 0; j < real_dim_count; j++) {
             blocks[i].idx[j] = blocks[i].start[j];
         }
     }
